First commit, lots of stuff. Wandering h's and Z's.
So, there's a flat world that a random number of zombies (Z) and
humans (h) spawn on. They wander around to random waypoints, and
a human next to a zombie gets turned and gets a new objective. The
Console class provides an interface to the terminal, and the Game
object ties it all together. `bin/run` makes the magic happen by
creating an instance of ZedSim::Game. Maps are made up of Tiles,
which are of a type, have an elevation, can contain multiple items
but only one Creature. That's about it.