Humans now run from zombies and use LOS to boot.
Human logic's pretty easy to consider: every tick they count the
zombies they can see. If there's a direction with less zombies,
they take it, otherwise they pick a "good" path at random (out of
the ones that are the safest, that is). I've also started to add
the back-end for colors, but it's not implemented yet.