Revamed human AI again, melee weapons are doneish
There are only machetes for now, but humans can find them and use
them to kill zombies. They also have an AI that uses various
priorities, such as finding a weapon or a group or just staying
alive. It's very extendable, and I expect to add zombie hunting
and other things to that list soon.